pyproject.toml
[project]
name = "dpd-pali-courses"
version = "0.1.0"
description = "Pāḷi course materials and website"
requires-python = ">=3.12"
dependencies = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "rich>=13.0.0",
    "PyYAML>=6.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "weasyprint>=68.1",
    "markdown>=3.10.2",
    "beautifulsoup4>=4.14.3",
    "pypandoc>=1.17",
    "python-docx>=1.2.0",
    "pdfminer-six>=20260107",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tools", "scripts"]
